index file reorganization
