+++ src/indices/future_index.h
@@ -5,44 +5,22 @@
 #ifndef BITCOIN_FUTUREINDEX_H
 #define BITCOIN_FUTUREINDEX_H
 
-#include <uint256.h>
 #include <amount.h>
+#include <indices/index.h>
 #include <script/script.h>
 #include <serialize.h>
-
-struct IndexKey {
-	uint256 txid;
-	unsigned int outputIndex;
-
-	ADD_SERIALIZE_METHODS;
-
-	template <typename Stream, typename Operation>
-	inline void SerializationOp(Stream& s, Operation ser_action) {
-		READWRITE(txid);
-		READWRITE(outputIndex);
-	}
-
-	IndexKey(uint256 hash, unsigned int index) {
-		txid = hash;
-		outputIndex = index;
-	}
-
-	IndexKey() {
-		SetNull();
-	}
-
-	void SetNull() {
-		txid.SetNull();
-		outputIndex = 0;
-	}
-};
+#include <uint256.h>
 
 struct CFutureIndexKey : IndexKey {
-	CFutureIndexKey(uint256 hash, unsigned int index) : IndexKey(hash, index) {
-	}
+    CFutureIndexKey(uint256 hash, unsigned int index) :
+        IndexKey(hash, index)
+    {
+    }
 
-	CFutureIndexKey() : IndexKey() {
-	}
+    CFutureIndexKey() :
+        IndexKey()
+    {
+    }
 };
 
 struct CFutureIndexValue {
@@ -56,7 +34,8 @@ struct CFutureIndexValue {
     ADD_SERIALIZE_METHODS;
 
     template <typename Stream, typename Operation>
-    inline void SerializationOp(Stream& s, Operation ser_action) {
+    inline void SerializationOp(Stream& s, Operation ser_action)
+    {
         READWRITE(satoshis);
         READWRITE(addressType);
         READWRITE(addressHash);
@@ -65,21 +44,24 @@ struct CFutureIndexValue {
         READWRITE(lockedToTime);
     }
 
-    CFutureIndexValue(CAmount amount, int type, uint160 addrHash, int confirmedBlock, uint32_t toHeight, int64_t toTime) {
-        satoshis = amount;
-        addressType = type;
-        confirmedHeight = confirmedBlock;
-        lockedToHeight = toHeight;
-        lockedToTime = toTime;
-        addressHash = addrHash;
+    CFutureIndexValue(CAmount amount, int type, uint160 addrHash, int confirmedBlock, uint32_t toHeight, int64_t toTime) :
+        satoshis(amount),
+        addressType(type),
+        addressHash(addrHash),
+        confirmedHeight(confirmedBlock),
+        lockedToHeight(toHeight),
+        lockedToTime(toTime)
+    {
     }
 
-    CFutureIndexValue() {
+    CFutureIndexValue()
+    {
         SetNull();
     }
 
-    void SetNull() {
-    	confirmedHeight = 0;
+    void SetNull()
+    {
+        confirmedHeight = 0;
         lockedToHeight = 0;
         lockedToTime = 0;
         satoshis = 0;
@@ -87,14 +69,15 @@ struct CFutureIndexValue {
         addressHash.SetNull();
     }
 
-    bool IsNull() const {
+    bool IsNull() const
+    {
         return addressHash.IsNull();
     }
 };
 
-struct IndexKeyCompare
-{
-    bool operator()(const IndexKey& a, const IndexKey& b) const {
+struct IndexKeyCompare {
+    bool operator()(const IndexKey& a, const IndexKey& b) const
+    {
         if (a.txid == b.txid) {
             return a.outputIndex < b.outputIndex;
         } else {
@@ -103,8 +86,7 @@ struct IndexKeyCompare
     }
 };
 
-struct CFutureIndexTxInfo
-{
+struct CFutureIndexTxInfo {
     std::map<CFutureIndexKey, CFutureIndexValue, IndexKeyCompare> mSpentInfo;
 };
