+++ src/indices/spentindex.h
